# docs/04-接口文档-openapi.yaml
openapi: 3.0.3
info:
  title: 供应链统筹任务管理平台 API
  version: 1.0.0
  description: 任务管理、配置管理、报表统计、消息推送接口
servers:
  - url: /api/v1/supply/task
paths:
...
  /tasks/{id}:
    get:
      summary: 查询任务详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTask'
  /tasks/{id}/transfer:
    post:
      summary: 转派任务
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetUserId:
                  type: string
      responses:
        '200':
          description: 成功
  /tasks/{id}/close:
    post:
      summary: 关闭任务
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: 成功
...
components:
  schemas:
...
    ApiResponseTask:
      type: object
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/TaskVO'
        message:
          type: string
        traceId:
          type: string
...
    TaskVO:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        status:
          type: string
        title:
          type: string
        productModel:
          type: string
        requester:
          type: string
        coordinator:
          type: string
        dueAt:
          type: string
        priority:
          type: integer
        createdAt:
          type: string